+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Journal des modifications
 - Interface de gestion des comptes locaux pour les présidents de jury.
 - Test d'archivabilité : possibilité de configurer le passage par un proxy.
 - Correction de l'import forcé de thèse : utilisation du unicaen/db-import pour synchroniser après l'import des données.
+- Ajout d'un filtre pour l'index des propositions de soutenance (index-structure)
 
 2.1.4
 -----